refactor: remove unused user actions and clean up update transformation page
