+++ app/(root)/transformations/[id]/update/page.tsx
@@ -1,13 +1,11 @@
-import { useRouter } from 'next/router'
+
 import React from 'react'
 
 const UpdateTransformationPage = () => {
-    const router = useRouter();
-    const { id } = router.query;
 
   return (
     <div>
-      UpdateTransformationPage {id}
+      UpdateTransformationPage 
     </div>
   )
 }
